package/hw-c910: Fixup pmp lock bit
We must set lock bit to prevent m speculitive access wrong address.
Reference 3.6.1

Locking and Privilege Mode
The L bit indicates that the PMP entry is locked, i.e., writes to the configuration register and
associated address registers are ignored. Locked PMP entries may only be unlocked with a system
reset. If PMP entry i is locked, writes to pmpicfg and pmpaddri are ignored. Additionally, if
pmpicfg.A is set to TOR, writes to pmpaddri-1 are ignored.

In addition to locking the PMP entry, the L bit indicates whether the R/W/X permissions are
enforced on M-mode accesses. When the L bit is set, these permissions are enforced for all privilege
modes. When the L bit is clear, any M-mode access matching the PMP entry will succeed; the
                ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
R/W/X permissions apply only to S and U modes.

Signed-off-by: Guo Ren <ren_guo@c-sky.com>